+++ packages/camera/camera_android_camerax/lib/src/process_camera_provider.dart
@@ -93,8 +93,8 @@ class ProcessCameraProviderHostApiImpl extends ProcessCameraProviderHostApi {
   /// Retrieves an instance of a ProcessCameraProvider from the context of
   /// the FlutterActivity.
   Future<ProcessCameraProvider> getInstancefromInstances() async {
-    return instanceManager.getInstanceWithWeakReference(await getInstance())!
-        as ProcessCameraProvider;
+    return instanceManager.getInstanceWithWeakReference<ProcessCameraProvider>(
+        await getInstance())!;
   }
 
   /// Gets identifier that the [instanceManager] has set for
@@ -114,7 +114,7 @@ class ProcessCameraProviderHostApiImpl extends ProcessCameraProviderHostApi {
     final List<int?> cameraInfos = await getAvailableCameraInfos(identifier);
     return cameraInfos
         .map<CameraInfo>((int? id) =>
-            instanceManager.getInstanceWithWeakReference(id!)! as CameraInfo)
+            instanceManager.getInstanceWithWeakReference<CameraInfo>(id!)!)
         .toList();
   }
 
@@ -138,8 +138,24 @@ class ProcessCameraProviderHostApiImpl extends ProcessCameraProviderHostApi {
       instanceManager.getIdentifier(cameraSelector)!,
       useCaseIds,
     );
-    return instanceManager.getInstanceWithWeakReference(cameraIdentifier)!
-        as Camera;
+    return instanceManager
+        .getInstanceWithWeakReference<Camera>(cameraIdentifier)!;
+  }
+
+  /// Returns whether or not the specified [UseCase] has been bound to the
+  /// lifecycle of the camera that this instance tracks.
+  Future<bool> isBoundFromInstances(
+    ProcessCameraProvider instance,
+    UseCase useCase,
+  ) async {
+    final int identifier = getProcessCameraProviderIdentifier(instance);
+    final int? useCaseId = instanceManager.getIdentifier(useCase);
+
+    assert(useCaseId != null,
+        'UseCase must have been created in order for this check to be valid.');
+
+    final bool useCaseIsBound = await isBound(identifier, useCaseId!);
+    return useCaseIsBound;
   }
 
   /// Returns whether or not the specified [UseCase] has been bound to the