+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -72,6 +72,14 @@ class ImageInformation {
   List<ImagePlaneInformation?> imagePlanesInformation;
 }
 
+@HostApi(dartHostTestHandler: 'TestInstanceManagerHostApi')
+abstract class InstanceManagerHostApi {
+  /// Clear the native `InstanceManager`.
+  ///
+  /// This is typically only used after a hot restart.
+  void clear();
+}
+
 @HostApi(dartHostTestHandler: 'TestJavaObjectHostApi')
 abstract class JavaObjectHostApi {
   void dispose(int identifier);