+++ packages/camera/camera_android_camerax/test/camera_selector_test.mocks.dart
@@ -58,3 +58,22 @@ class MockTestCameraSelectorHostApi extends _i1.Mock
         returnValue: <int?>[],
       ) as List<int?>);
 }
+
+/// A class which mocks [TestInstanceManagerHostApi].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockTestInstanceManagerHostApi extends _i1.Mock
+    implements _i2.TestInstanceManagerHostApi {
+  MockTestInstanceManagerHostApi() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  void clear() => super.noSuchMethod(
+        Invocation.method(
+          #clear,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+}