+++ packages/camera/camera_android_camerax/test/image_capture_test.dart
@@ -76,7 +76,7 @@ void main() {
       instanceManager.addHostCreatedInstance(
         imageCapture,
         0,
-        onCopy: (_) => ImageCapture.detached(),
+        onCopy: (_) => ImageCapture.detached(instanceManager: instanceManager),
       );
 
       imageCapture.setFlashMode(flashMode);