+++ packages/google_maps_flutter/google_maps_flutter/example/integration_test/google_maps_test.dart
@@ -469,11 +469,15 @@ void main() {
       expect(
           coordinate.x,
           ((rect.center.dx - rect.topLeft.dx) *
+                  // TODO(pdblasi-google): Update `window` usages to new API after 3.9.0 is in stable. https://github.com/flutter/flutter/issues/122912
+                  // ignore: deprecated_member_use
                   tester.binding.window.devicePixelRatio)
               .round());
       expect(
           coordinate.y,
           ((rect.center.dy - rect.topLeft.dy) *
+                  // TODO(pdblasi-google): Update `window` usages to new API after 3.9.0 is in stable. https://github.com/flutter/flutter/issues/122912
+                  // ignore: deprecated_member_use
                   tester.binding.window.devicePixelRatio)
               .round());
     }