+++ packages/google_maps_flutter/google_maps_flutter_android/android/src/test/java/io/flutter/plugins/googlemaps/GoogleMapControllerTest.java
@@ -14,6 +14,7 @@ import static org.mockito.Mockito.verify;
 
 import android.content.Context;
 import android.os.Build;
+import android.os.Looper;
 import androidx.activity.ComponentActivity;
 import androidx.test.core.app.ApplicationProvider;
 import com.google.android.gms.maps.GoogleMap;
@@ -22,6 +23,7 @@ import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import java.util.HashMap;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -31,6 +33,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
+import org.robolectric.Shadows;
 import org.robolectric.annotation.Config;
 
 @RunWith(RobolectricTestRunner.class)
@@ -41,17 +44,30 @@ public class GoogleMapControllerTest {
   private ComponentActivity activity;
   private GoogleMapController googleMapController;
 
+  AutoCloseable mockCloseable;
   @Mock BinaryMessenger mockMessenger;
   @Mock GoogleMap mockGoogleMap;
 
   @Before
   public void before() {
-    MockitoAnnotations.initMocks(this);
+    mockCloseable = MockitoAnnotations.openMocks(this);
     context = ApplicationProvider.getApplicationContext();
-    activity = Robolectric.setupActivity(ComponentActivity.class);
+    setUpActivityLegacy();
     googleMapController =
         new GoogleMapController(0, context, mockMessenger, activity::getLifecycle, null);
     googleMapController.init();
+  }
+
+  // TODO(stuartmorgan): Update this to a non-deprecated test API.
+  // See https://github.com/flutter/flutter/issues/122102
+  @SuppressWarnings("deprecation")
+  private void setUpActivityLegacy() {
+    activity = Robolectric.setupActivity(ComponentActivity.class);
+  }
+
+  @After
+  public void tearDown() throws Exception {
+    mockCloseable.close();
   }
 
   @Test
@@ -104,6 +120,7 @@ public class GoogleMapControllerTest {
 
       verify(mapView, never()).invalidate();
       argument.getValue().onMapLoaded();
+      Shadows.shadowOf(Looper.getMainLooper()).idle();
       verify(mapView).invalidate();
     }
   }
@@ -127,6 +144,7 @@ public class GoogleMapControllerTest {
 
     verify(mapView, never()).invalidate();
     argument.getValue().onMapLoaded();
+    Shadows.shadowOf(Looper.getMainLooper()).idle();
     verify(mapView).invalidate();
   }