+++ packages/google_maps_flutter/google_maps_flutter_android/example/integration_test/google_maps_tests.dart
@@ -471,11 +471,15 @@ void googleMapsTests() {
     expect(
         coordinate.x,
         ((rect.center.dx - rect.topLeft.dx) *
+                // TODO(pdblasi-google): Update `window` usages to new API after 3.9.0 is in stable. https://github.com/flutter/flutter/issues/122912
+                // ignore: deprecated_member_use
                 tester.binding.window.devicePixelRatio)
             .round());
     expect(
         coordinate.y,
         ((rect.center.dy - rect.topLeft.dy) *
+                // TODO(pdblasi-google): Update `window` usages to new API after 3.9.0 is in stable. https://github.com/flutter/flutter/issues/122912
+                // ignore: deprecated_member_use
                 tester.binding.window.devicePixelRatio)
             .round());
     await tester.binding.setSurfaceSize(null);