+++ packages/path_provider/path_provider_windows/example/lib/main.dart
@@ -13,7 +13,7 @@ void main() {
 
 /// Sample app
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   State<MyApp> createState() => _MyAppState();