+++ packages/webview_flutter/webview_flutter_android/lib/src/android_proxy.dart
@@ -40,8 +40,7 @@ class AndroidWebViewProxy {
     Future<List<String>> Function(
       android_webview.WebView webView,
       android_webview.FileChooserParams params,
-    )?
-        onShowFileChooser,
+    )? onShowFileChooser,
   }) createAndroidWebChromeClient;
 
   /// Constructs a [android_webview.WebViewClient].
@@ -52,21 +51,18 @@ class AndroidWebViewProxy {
       android_webview.WebView webView,
       android_webview.WebResourceRequest request,
       android_webview.WebResourceError error,
-    )?
-        onReceivedRequestError,
+    )? onReceivedRequestError,
     @Deprecated('Only called on Android version < 23.')
         void Function(
       android_webview.WebView webView,
       int errorCode,
       String description,
       String failingUrl,
-    )?
-            onReceivedError,
+    )? onReceivedError,
     void Function(
       android_webview.WebView webView,
       android_webview.WebResourceRequest request,
-    )?
-        requestLoading,
+    )? requestLoading,
     void Function(android_webview.WebView webView, String url)? urlLoading,
   }) createAndroidWebViewClient;
 
@@ -88,8 +84,7 @@ class AndroidWebViewProxy {
       String contentDisposition,
       String mimetype,
       int contentLength,
-    )
-        onDownloadStart,
+    ) onDownloadStart,
   }) createDownloadListener;
 
   /// Enables debugging of web contents (HTML / CSS / JavaScript) loaded into any WebViews of this application.