+++ script/tool/test/lint_android_command_test.dart
@@ -39,6 +39,40 @@ void main() {
       runner.addCommand(command);
     });
 
+    void writeFakeBuildGradle(RepositoryPackage example, String pluginName,
+        {bool warningsConfigured = true}) {
+      final String warningConfig = '''
+gradle.projectsEvaluated {
+    project(":$pluginName") {
+        tasks.withType(JavaCompile) {
+            options.compilerArgs << "-Xlint:all" << "-Werror"
+        }
+    }
+}
+''';
+      example
+          .platformDirectory(FlutterPlatform.android)
+          .childFile('build.gradle')
+          .writeAsStringSync('''
+buildscript {
+    repositories {
+        google()
+        mavenCentral()
+    }
+    dependencies {
+        classpath 'com.android.tools.build:gradle:8.0.1'
+    }
+}
+allprojects {
+    repositories {
+        google()
+        mavenCentral()
+    }
+}
+${warningsConfigured ? warningConfig : ''}
+''');
+    }
+
     test('runs gradle lint', () async {
       final RepositoryPackage plugin =
           createFakePlugin('plugin1', packagesDir, extraFiles: <String>[
@@ -46,6 +80,7 @@ void main() {
       ], platformSupport: <String, PlatformDetails>{
         platformAndroid: const PlatformDetails(PlatformSupport.inline)
       });
+      writeFakeBuildGradle(plugin.getExamples().first, 'plugin1');
 
       final Directory androidDir =
           plugin.getExamples().first.platformDirectory(FlutterPlatform.android);
@@ -83,6 +118,9 @@ void main() {
           platformSupport: <String, PlatformDetails>{
             platformAndroid: const PlatformDetails(PlatformSupport.inline)
           });
+      for (final RepositoryPackage example in plugin.getExamples()) {
+        writeFakeBuildGradle(example, 'plugin1');
+      }
 
       final Iterable<Directory> exampleAndroidDirs = plugin.getExamples().map(
           (RepositoryPackage example) =>
@@ -140,6 +178,7 @@ void main() {
       ], platformSupport: <String, PlatformDetails>{
         platformAndroid: const PlatformDetails(PlatformSupport.inline)
       });
+      writeFakeBuildGradle(plugin.getExamples().first, 'plugin1');
 
       final String gradlewPath = plugin
           .getExamples()
@@ -162,6 +201,34 @@ void main() {
           output,
           containsAllInOrder(
             <Matcher>[
+              contains('The following packages had errors:'),
+            ],
+          ));
+    });
+
+    test('fails if javac lint-warnings-as-errors is missing', () async {
+      final RepositoryPackage plugin =
+          createFakePlugin('plugin1', packagesDir, extraFiles: <String>[
+        'example/android/gradlew',
+      ], platformSupport: <String, PlatformDetails>{
+        platformAndroid: const PlatformDetails(PlatformSupport.inline)
+      });
+      writeFakeBuildGradle(plugin.getExamples().first, 'plugin1',
+          warningsConfigured: false);
+
+      Error? commandError;
+      final List<String> output = await runCapturingPrint(
+          runner, <String>['lint-android'], errorHandler: (Error e) {
+        commandError = e;
+      });
+
+      expect(commandError, isA<ToolExit>());
+      expect(
+          output,
+          containsAllInOrder(
+            <Matcher>[
+              contains('The example example is not configured to treat javac '
+                  'lints and warnings as errors.'),
               contains('The following packages had errors:'),
             ],
           ));