+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -12,4 +12,7 @@
 * Adds implementation of availableCameras().
 * Implements camera preview, createCamera, initializeCamera, onCameraError, onDeviceOrientationChanged, and onCameraInitialized.
 * Adds integration test to plugin.
+* Bump CameraX version to 1.3.0-alpha04.
+* Fixes instance manager hot restart behavior and fixes Java casting issue.
 * Implements image capture.
+* Implements image streaming.