+++ packages/camera/camera_avfoundation/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.9.13+1
+
+* Clarifies explanation of endorsement in README.
+
+## 0.9.13
+
+* Allows camera to be switched while video recording.
+* Aligns Dart and Flutter SDK constraints.
+
 ## 0.9.12
 
 * Updates minimum Flutter version to 3.3 and iOS 11.