+++ packages/file_selector/file_selector_ios/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.1+1
+
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
+## 0.5.1
+
+* Updates minimum Flutter version to 3.3 and iOS 11.
+
 ## 0.5.0+3
 
 * Updates links for the merge of flutter/plugins into flutter/packages.