+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.8.7
+
+* Adds `usePhotoPickerAndroid` options.
+* Aligns Dart and Flutter SDK constraints.
+
 ## 0.8.6+4
 
 * Updates iOS minimum version in README.