+++ packages/path_provider/path_provider/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.14
+
+* Updates README to use code excerpts.
+* Aligns Dart and Flutter SDK constraints.
+
 ## 2.0.13
 
 * Updates links for the merge of flutter/plugins into flutter/packages.