+++ packages/path_provider/path_provider_foundation/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.2.0
+
+* Adds support for `getContainerPath` on iOS.
+
+## 2.1.3
+
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
 ## 2.1.2
 
 * Updates links for the merge of flutter/plugins into flutter/packages.