+++ packages/url_launcher/url_launcher_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 6.0.26
+
+* Bump RoboElectric dependency to 4.4.1 to support AndroidX.
+
+## 6.0.25
+
+* Clarifies explanation of endorsement in README.
+* Aligns Dart and Flutter SDK constraints.
+
 ## 6.0.24
 
 * Updates links for the merge of flutter/plugins into flutter/packages.